requirements.txt
# Agora Discord Bot Requirements

# Discord.py - Main Discord library
discord.py>=2.3.0

# Database libraries
SQLAlchemy>=2.0.0
asyncpg>=0.29.0  # PostgreSQL async driver
redis>=5.0.0     # Redis client with async support

# HTTP and API libraries
aiohttp>=3.8.0   # For LLM API calls
requests>=2.31.0 # Fallback HTTP client

# Image processing (for emoji validation)
Pillow>=10.0.0

# Environment and configuration
python-dotenv>=1.0.0

# System monitoring
psutil>=5.9.0

# Development and debugging
psutil>=5.9.0    # For system monitoring in status command

# Async utilities
asyncio-mqtt>=0.13.0  # If needed for future message queuing